fix playlist

+++ src/main/java/it/unical/sadstudents/mediaplayeruid/view/ContextMenuHandler.java
@@ -33,15 +33,15 @@ public class ContextMenuHandler extends ContextMenu {
             menuItem1.setOnAction(new EventHandler<ActionEvent>() {
                 @Override
                 public void handle(ActionEvent actionEvent) {
-                    if(SceneHandler.getInstance().showConfirmationAlert("Do you really want to delete the file?")) {
-                        if (source == "home") {
+
+                        if (source == "home"  && SceneHandler.getInstance().showConfirmationAlert("Do you want to remove from Recent Media?")) {
                             for (int i = 0; i < Home.getInstance().getRecentMedia().size(); i++) {
                                 if (myMedia.equals(Home.getInstance().getRecentMedia().get(i))) {
                                     Home.getInstance().removeItem(i);
                                     break;
                                 }
                             }
-                        } else if(source == "video"){
+                        } else if(source == "video" && SceneHandler.getInstance().showConfirmationAlert("Do you really want to delete this Video?")){
                             for (int i = 0; i < VideoLibrary.getInstance().getVideoLibrary().size(); i++) {
                                 if (myMedia.equals(VideoLibrary.getInstance().getVideoLibrary().get(i))) {
 
@@ -58,7 +58,7 @@ public class ContextMenuHandler extends ContextMenu {
                                 }
                             }
                         }
-                        else if(source == "musicLibrary"){
+                        else if(source == "musicLibrary" && SceneHandler.getInstance().showConfirmationAlert("Do you really want to delete this Song?") ){
                             for (int i=0; i<MusicLibrary.getInstance().getMusicLibrary().size(); i++){
                                 if(myMedia.equals(MusicLibrary.getInstance().getMusicLibrary().get(i))){
                                     MusicLibrary.getInstance().deleteWithIndex(i);
@@ -68,7 +68,7 @@ public class ContextMenuHandler extends ContextMenu {
 
                         }
 
-                    }
+
                 }
             });
